+++ src/renderer/utils/exportSetting.ts
@@ -1,6 +1,7 @@
 export const createExportSetting = (setting: LX.AppSetting): LX.AppSetting => {
   const data = JSON.parse(JSON.stringify(setting))
   data['network.gitcodeMusicAccessToken'] = ''
+  data['sync.webdav.password'] = ''
   return data
 }